pyproject.toml
[project]
name = "maestro-ai-server"
version = "0.1.0"
description = "Maestro AI Service - 基于 LangChain 的 AI 断言和文本提取服务"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "LJY" }
]
dependencies = [
    # Web 框架
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.18",
    # LangChain 核心
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langsmith>=0.2.0",
    # 数据验证
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    # 可靠性
    "tenacity>=9.0.0",
    # 日志
    "structlog>=24.0.0",
    # 图像处理
    "pillow>=11.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "pytest-cov>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
